backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
constantly==23.10.4
cryptography==44.0.0
daphne==4.1.2
defusedxml==0.8.0rc2
Django==5.1.3
django-cors-headers==4.6.0
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.0
oauthlib==3.2.2
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-nmap==0.7.1
python3-openid==3.2.0
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
scapy==2.6.1
service-identity==24.2.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
zope.interface==7.2
